zeropoint: First make file for obtainaing the zeropoint photometry
Until now, we did not have any scripts for obtaing the zeropoint in
Gnuastro.

With this commit, based on any number of reference images which have a
zeropoint, we can calculate the zeropoint of the input image by using this
Makefile.

This job is not yet complete: this Makefile will be called from a script
that users should call. But this script has not been written, yet.

This Makefile is a rewrite of a similar Makefile of Samane Raji. Also,
Mohammad Akhlaghi helped in editing and debugging.