zeropoint: catalogs can be used as reference
Until now, in the 'zeropoint.mk' file if the user had an image as input and
one (or more than one) image as refrence, s/he could obtain the zeropoint
of the input image based on the zeropoint of the refrence image. But if the
user had an image as input and a catalog as refrence and wanted calculate
the zeropoint of the image, this option did not exist in the Make file.

With this commit, this option is added to the Makefile. Now user can based
on the zeropoint of a catalog (as a refrence) can calculate the zeropoint
of the input image.