Zeropoint: script for estimating the zeropoint of an image
Until now, the make file for estimating the zeropoint of an image from a
catalog or image(s) was ready. But if want to bring it in Gnuastro, the
Makefile should be called inside the script.

With this commit, the configuration file of the Makefile is prepared
automatically, based on the options of zeropoint's script and finally the
zeropoint of the image from the catalog or image(s) is estimated.