Book: more general link to DS9's binary download in itemized list
Until now, for the installation of DS9 (in an appendix) we were guiding the
readers to the top Google-based page of DS9. However, that page has many
limiations (for example it does not open in some countries!) and generally
has a terrible interface making it hard to find the download link.

With this commit, the direct download page https://ds9.si.edu/download is
now given and we have a simple checklist for the steps to check the
downloaded file and to install it.

This commit and its message were edited by Mohammad Akhlaghi.