Zeropoint: keep Mag-vs-Zeropoint for best aperture in output
Until now, if the user did not use the '--keepzpap' option best aperture
and its zeropoint just was written in header of output. Although, for input
image we did not consider any value for the input Zeropoint.

With this commit, even if user do not use the '--keepzpap' option the
Magnitude versus Zeropoint table is saved in the second extension of the
output.  On the other hand, in making the catalog Zeropoint of the input
image consider as zero.