zeropoint: use --magrage provided by user
Until now the magnitdue range was hard-coded in the script and not possible
to change at run-time. Also, the range was not written in the output FITS
file.

With this commit, the value of the '--magrange' option (set at run-time) is
now used and the value is written in the output FITS file as a keyword.