Zeropoint: add magnitude range into the header of output
Until now, if user considered the magnitude range did not write in header
of output.

With this commit, two values named 'MAGMIN' and 'MAGMAX' are added to
header of output. If user determine the magnitude range both of these will
take value, otherwise if user do not determine magnitude range both of them
do not have value in header.