Library (type.h): using %e for floats smaller than precision
Until now, in 'gal_type_to_string' we used '%.6f' and '%.14f' for 32-bit
and 64-bit floats respectively. This would cause in numbers that are
smaller than those powers to be printed as zero! For example 1e-8 as a
32-bit floats would be printed as '0.000000'!

With this commit, when the number to convert to a string of characters is
less than 1.0, this function will use '%e', otherwise, it will use '%f'
like before.

This bug was reported by Elham Saremi.

This fixes bug #63340.