Book: tutorial of zero point; preparing images for using the script
Until now, I wrote an introduction for this tutorial.

With this commit, I started to write the first subsection named "Using the
image as a zero point reference". In this step, I selected two images of
SDSS as a reference and an image of J-PLUS to find its zero
point. Downloading and preparing of data in this step are written.