Library (type.h/table.h): default printing of floats in exp mode
Until now, when any of the programs were asked to build a plain-text table
they would do so in fixed-point notation. But this could cause a loss of
information/precision (especially when the column contains values with a
very large dynamic range).

With this commit, the default mode for printing floating points (as single
numbers or as table columns) has been changed to exponential mode. This
fixes the problem of potential loss of information. Also to help clarify
the issue to a new user, a new sub-section has been added under the Table
program to describe all the intricacies with printing floating point
numbers and how to best manage their columns in the respective
types. Furthermore, the counting of significant digits when finding the
type of a floating point string input has also been corrected (previously
it wouldn't properly count the digits in an exponential notation).