Book: tutorial of zero point; select the best aperture
Until now, I explained how users can select a proper range of
magnitude.

With this commit, I wrote a part related to choosing the best
aperture size for finding a more accurate zero point and added
a table for results.

I decided to don't use a plot in the book for showing the range
of magnitude but ask to the user to use the TOPCAT for drawing
the plot.

Also, I referred to the importance of sky subtracting in the
results.