Library (arithmetic): new operators to convert counts to nanomaggies
Until now, there was no easy way to convert the counts of an image to
nanomaggies (used by SDSS and many others, with a fixed zeropoint of 22.5)
or vice-versa.

With this commit, the necessary library functions have been written and the
interface to the 'arithmetic.h' library has been added to allow this
operation easily in both the Arithmetic and Table programs.

This was suggested by Giulia Golini.