Zeropoint: adding sanity check for aperture option
Until now, we did not have a sanity check for --aperarcsec option. Without
this, the user did not understand that the existence of it is necessary.

With this commit, a sanity check is added to zero point script. So if the
user doesn't use this option, one error will be printed.