Zeropoint: a bug about the magnitude range
Until now, if the user used the option --magnituderange with two values
minimum and maximum of magnitude, there was no problem. But if the user
didn't determine this range, an error is encountered. While this option is
not mandatory.

With this commit, the bug is fixed. The bug was related to setting minimum
and maximum magnitude on the header of output while there was no condition
for when users don't use the range.