Book: Copyright added to library demo programs
Until now, there was no copyright for programs of library demo and the text
of the book needed some modification to better understand.  Also, There
were a few typos like "tar-ball" or "re-sample" instead of "tarball" or
"resample".

With this commit, a copyright statement has been added with a short
description of the program above it, to follow good practices in to the
library demo programs. The typos have also been corrected.