Library (arithmetic.h): alt string names for type conversion operators
Until now, like all the operators of Arithmetic, the type conversion
operators of Arithmetic only had a single format. However, when writing
column metadata in plain-text, we allow for two different formats (for
example 'i16' or 'int16'). The fact that the short formats weren't
available in Arithmetic was annoying.

With this commit, the same short strings can now be used to identify the
type conversion operators of Arithmetic (a single operator can be called
with two different names on the command-line).

This was added after a discussion with Samane Raji.