Library (blank.h): gal_blank_flag returns uint8 output on empty input
Until now, when the input to 'gal_blank_flag' was an empty dataset, the it
would simply return the input! However, this would ignore the fact that the
input can have any data type, and doesn't have to be an unsigned 8-bit
integer (which is the expected format of a flag).

As a result, when the Table program was called with '--noblank' or
'--noblankend' on an empty input table, the Table program would abort,
saying that the type of the flag is not correct! Also, when '--noblank' was
used on an empty table, Table would give a segmentation fault!

With this commit, when the input to 'gal_blank_flag' is empty, a new empty
dataset is allocated with the unsigned 8-bit integer type and returned as
flag. This fixed the problem with the wrong type of flag. For the
segmentation fault, I noticed that we are only checking for 'dsize' being
NULL (to identify if the table is empty!), but other scenarios may also
happen. So we now also check the size and the presense of an array.

This bug was reported by Elham Saremi.

This fixes bug #63399.