Book: editing zero point estimation section
Until now, zero point estimation section was not complete and obtaining the
zero point based on the reference images was not edditing.

With this commit, zero point estimation section is written completely and
some parameters are added to the invoking section. Finally, obtaining the
zero point based on the references images is edited.