Book: early data release 3 of Gaia change to the data release 3
Until now, almost all the usages of the Gaia dataset with Query in the
Gnuastro book pointed to the early datarelease 3 dataset. However, Gnuastro
now also enables querying from data release 3.

With this commit, the early data release 3 usages in the book (and in the
'astscript-psf-select-stars' script) have been changed to the DR3.