Book: tutorial of zero point; subsection catalog reference
Until now, I wrote "Photometric calibration of images by zero point" and
"Zero point based on the reference image" sections.

With this commit, I wrote a part of "Zero point based on the reference
catalog" section. In this section, I used a SDSS catalog overlapped with
the J-plus image to estimate the zero point of the image based on the SDSS
catalog.