MakeCatalog: added --sigclip-mean-sb for radius independent SB
Until now, to only easy way to estimate the surface brightness profile in
the radial profile script was to use the '--surfacebrightness' measure of
MakeCatalog. However, the surface brightness measure of MakeCatalog is done
over the full area of the label (circumference in the case of a radial
profile). And the area of the circumference where the radial profile is
measured depends on the radius! Therefore the vertical axis will not be
independent of the horizontal, making it hard to interpret and technically
wrong.

With this commit, a new column has been added to MakeCatalog for such
scenarios: '--sigclip-mean-sb'. This will estimate a single per-pixel value
for each label (radius in a radial profile) and estimate the surface
brightness over the pixel area. This makes the surface brightness profile
independent of the radius, and physically interpretable.

In the process, I also noticed that our usage of "--brightness" for counts
in MakeCatalog, MakeProfiles and MakeNoise was confusing (with the physical
meaning of "brightness" as energy/time). Therefore the options that had
"brightness" in their name were changed to something that is more
clear. For example MakeCatalog's '--brightness' column is now called
'--sum', see the "Changed features" of the NEWS file for more.

In the process, I also noticed a Texinfo warning about the usage of '.' in
the node of the GNU Free Documentation License within the book. Therefore
the '.' has been removed from the node name.

This was done after a discussion raised by Ignacio Trujillo and Zahra
Sharbaf.