Zeropoint: a bug about the matching of reference and main catalog
Until now, to match objects in the main and reference catalog, labels were
used. Because IDs were the same for both catalogs when there were two
images and both catalog were obtained with aperture photometry.
But when there is a prepared catalog as a reference, so it is needed to do
a match based on RA and DEC.

With this commit, the bug is solved. Also, some of the typos in zeropoint
script are corrected.