Book: open imge by 'zscale' scale and options are written as code format
Until now, for visualizing the SDSS catalogs object in J-PLUS image the
scale did not determine and it did not make sense. Also, options did not
write as code format.

With this commit, scale of imge is 'zcale'. Now if upload the region on the
imge by Ds9 it make sense. Option are wrote as code format.