developer-build: add the --confopts option
Until now, one could not set the 'configure' options while calling the
'developer-build' script. For example, one might not need to build a
specific feature, like those that are related to 'python'. Normally the
caller could use './configure --without-python' to achieve this. However,
there was no way of doing that while using the 'developer-build' script.

With this commit, the '--confopts' option is added to the 'developer-build'
script which takes a string sandwiched between double quotations so that
the 'configure' options would not get mixed up with the 'developer-build'
options. For instance, now one could install gnuastro while excluding
programs that need python like this:

    ./developer-build --confopts "--without-python" -I