Book: added backslash in command line-break that was missing
Until now, the alias command that we had defined for 3D NoiseChisel was
missing a backslash at the end of the broken line!

With this commit a backslash has been added. Also, the backslash of the
previous command which was too far away has been brough closer to the
visible end of the line.