Configuration: Python extensions have to be manually activated
Until now, Gnuastro would always attempt to build the Python extensions if
the host had the necessary Python headers. However, due to the many
peculiarities of Python, this caused some problems for special cases where
Python wasn't necessary at all.

With this commit, Python extensions will no longer be included by default!
Even if Python is present on the host, if the user doesn't configure
Gnuastro with the '--with-python' option, they will not be used.

This change was implemented after the feedback of Zohreh Ghaffari.