Book: added box under the + operator for cases that involve NaNs
Until now, the description of the '+' operator didn't mention how it treats
NaN values.

With this commit, a box has been added at the end of the description of the
'+' operator that describes the problem and solution (to use the 'sum'
operator of the stacking operators).

This issue was raised by Elham Saremi, and before this by Giulia Golini and
Raul Infante-Sainz.