radial-profile: new options for precision and radial interval
Until now, the radial profile script could only measure radii on an integer
basis. For example all the 8-connected neighbors of central pixel would be
considered at radius 2! Also, when oversampling was called, it wouldn't put
the central over-sampled pixel in the center of the original-resolution
pixel! Furthermore, when you wanted to build a 2D profile, it wasn't easy
to prepare the output for giving as a '--customtable' in MakeProfiles!

With this commit, all these issues are addressed: the new '--precision'
option allows for having a higher precision in sampling the profile, the
bug with oversampling has been fixed by accounting for this fact and a new
option called '--radinterval' has been added which will cause the output to
report intervals instead of a single value (ready to be fed to
MakeProfiles). This was particularly necessary after adding the precision
feature.

This bug was reported by Nafiseh Sedighi.

This fixes bug #63485.