Book: link to new --radinterval option added in MakeProfiles
Until now, the newly added option in the radial profile script that will
greatly help in running MakeProfiles with a custom profile was not
mentioned in the MakeProfiles section of the book!

With this commit, a small paragraph has been added in the description of
MakeProfile's '--customtable' that points the interested readers to this
feature in the radial profile script.