Table: new sorted-to-interval operator in column arithmetic
Until now, the radial profile script had the '--radinterval' option to
generate an input to the '--customtable' option of MakeProfile. It had
multiple problems:
  1. When the output was generated in this format, the raw profile was no
     longer created.
  2. It was hard to do some operations on the radial profile (like merging
     multiple profiles) and generate an input to the '--customtable' option
     of MakeProfiles.

With this commit, that option has been removed from the radial profile
script and instead, a new column arithmetic operator has been added that is
called 'sorted-to-interval'. This allows users to easily generate an input
to the '--customtable' option of MakeProfiles from any radial profile
easily. This fixes both the problems above and is much more modular and
generic.