Book: revision of the first part of the zeropoint tutorial
Until now, the first part of zeropoint tutorial (where an image
is used as a reference) was ok but the fluency of the text and
readability could be improved.

With this commit, I have revised the text, I changed some phrases
to improve the text and I have corrected few minor grammatical
errors.