Book: emphasis to determine the pixel scale of images in warping
Until now, there was no emphasis that users should use the '--cdelt' option
for fixing the pixel scales images when they want to stack the
outputs. This is important because the pixel sizes will be a little
different after astrometry which is a normal/statistical error. However,
for stacking it is necessary to we have exactly the same pixel sizes. So it
is necessary to fix a constant value for example up to two decimal places
with '--cdelt' option.

With this commit, this paragraph was added to the book.