All programs: error when input and output names are the same
Until now, when the given (or automatically selected) output file name of a
program was the same as the input, the program would delete the output file
(and thus the input) before reading it! As a result, it would delete the
input file and also complain that it doesn't exist. This would be a very
confusing error (and possibly even resulting in loss of data!).

With this commit, the function that checks the writability of the output
(and deletes it, if it exists), also takes the name of the program's input
file and will abort if the input or output names are the same.

This bug was reported by Faezeh Bidjarchian.

This bug was reported by bug #63543.