Book: Added explanations about the default HDU
Until now, there was no clear explanation about whether we have to specify
the HDU number for FITS files in our command or not. And if we don't
specify it, what the default value will be. Also there was not an clear
example in the “Filtering (smoothing) operators” subsection for
re-inserting NaN pixels to final output (since the filtering operator
affects the NaNs). In addition, a few typos were found.

With this commit, some explanations about using the HDU number in our
command have been added. Also, an example on the “Filtering (smoothing)
operators” subsection has been added. As well as some corrections have been
performed.