Book: clear description of the --publish in developer-build script
Until now, in the book, we hadn't explicitly mentioned that the '--publish'
option of the 'deveveloper-build' script needs a value.

With this commit, the description of this option has been edited to clarify
that it takes a string argument and the string is interpretted to be the
URL of the place you want to publish to.

This was reported by Faezeh Bidjarchian.