Library (arithmetic.h): box-around-ellipse accounts for floats
Until now, the 'box-around-ellipse' operator would assume that the inputs
are always double-precision floating point and would give segmentation
faults for other types!

With this commit, the condition that was causing this has been fixed. Also,
the description of this operator in the book has claried that the two input
lengths should be radius, not diameter for example.