Library (arithmetic.h): new box-vertices-on-sphere operator
Until now, in several places within Gnuastro we accounted for the fact that
a box's vertices on a sphere need special consideration. However, for an
outside user, doing the job manually would require a complex column
arithmetic operation (and to get to pen-and-paper to derive the
equations!).

With this commit, a simple arithmetic operator has been added for this job:
'box-vertices-on-sphere'. It will take the center and size of the box along
each dimension and will return the coordinates of the 8 corners, while
accounting for the spherical effects. Some real-world usage examples and a
description of the problem it is designed to solve have been given in the
extensive documentation of this new operator.

This was added after a discussion with Martin Kuemmel.