Table: when printing column info, also shows vector columns
Until now, Table (actually the 'gal_table_print_info' function in the
library) would not show if the given column is a vector (multi-value) or
not.

With this commit, when the column is a vector, a '[N]' is printed after the
column's type. The 'N' shows how many values there are in that vector.