Zeropoint: add an option for doing many jobs and check overlapping
Until now, we did not have any option for allowing script do many jobs at
the same time.  On the other hand, we did not check that input image and
refernce images and catlog overlap together or not.

With this commit, an option of `-j` or `--jobs` is added to the script to
allow the script do many things at the same time. By default the value of
this option is one and it do one job. If user wants to do many thing s/he
should change this number.
Finally, two condition is added to the script to check that input image and
references images and catalog overlap together or not. If they overlap the
pipeline will be run correctly. But if they do not overlap together the
pipeline will stop and print an error.