Zeropoint: optimize the script for overlaping check
Until now, if the reference was catalog it check that input image and
catalog overlap together or not. While, if the user has reference images
this criteria did not check.

With this commit, if the user has reference images and based on them wants
obtain the zero point, it will check they have overlap with input image or
not.