MakeCatalog: per-slice measurements in vector columns
Until now, MakeCatalog would write per-slice measurements (like the
spectrum of an object in an IFU cube) in different HDUs of the output
file. This was extremely slow and very inconvenient.

With this commit, MakeCatalog now writes such measurements as vector
columns in its single output table and the old '--spectrum' column has been
removed. It has been replaced by separate measurements with this format:
'--*-in-slice'.

The following changes have also been made as part of this commit:

 - Book: Added new tutorial to show-case this feature. Which is also a good
   demonstration of vector column management.

 - Library (spectrallines.h): added a much larger database of lines.

 - CosmicCalculator: new option to specify the line unit to input/output.