Book: some command and some tips is added to better recognition
Until now, just in first section was explained how open 'TOPCAT' and find
the best magnitude. Eventhough, we had a same paragraph two times.

With this commit, in each step complete command for how opening the
'TOPCAT' and compare all the results together is written. Although, some
commands for best recognition is added to the main body of book and extra
paragraph is removed.