astscript-zeropoint: new option --matchradius added
Until this commit, when matching with the reference catalog, a fixed
matching distance (0.2 arcsec) was considered. However, it is not good to
have such important parameter fixed. In some situations it would be
necessary to change this parameter.

With this commit, the possibility of changing this matching distance has
been added. Now the zero point script has an option, -s/--matchradius, to
modify that parameter.