astscript-zeropoint: writting all receipes as in one shell
Until this commit, in the Makefile for the zero point calculation, we were
assuming a GNU Make. However, in some systems this is not true. For
example, on macOS computers it is actually GNU Make but it is not able to
enable the .ONESHELL. target. Consequently, in order to make this script
portable, it is necessary to consider each receipe as one single shell.

With this commit, all lines within one receipe have been concatenated in
order to make them as one single line. I.e., they are now executed in one
single shell.  This makes the code harder to read and depurate but it is
more portable. Still it is necessary to improve the comments and
information above each receipe.