Book: simplifying and polishing the zero point script tutorial
Until this commit, the zero point script tutorial had very detailed
information that were actually redundant in some cases. In order to
simplify it and make it shorter for the reader, it was necessary to cut and
re-organize some parts.

With this commit, I modified several parts in order account for the above.
I remove information that is already exposed in other sections of Gnuastro
book so the reader can go directly there. In addition to that, I tried to
polish and correct typos in the text.