astscript-zeropoint: Content of temorary file have been removed
Until now, the content of the temporary directory were not considered and
if the user changed a parameter the final result did not change. Because
zero point value obtained in the make file and all the prerequiesit were in
temporary directory and due to this make did not build them again.
Although, in the mannual of the zeropoitn some commands did not start with
'$' sign.

With this commit, first of all in the script the all the contents of the
temporary directory will be removed and all files will created newly.  On
the other hand, in zero point mannual all the mannual beginig with '$'
sign.