Book: downloading dependency tarballs with wget command
Until now, readers needed to click on the URL to download the source
tarballs of three mandatory dependencies. Afterwards, they needed to go to
their download directory run the rest of the commands. Besides being
inconvenient, this has recently started to cause problems with
WCSLIB. Because WCSLIB's source is distributed in an FTP server and newer
versions of the browsers don't support FTP any more.

With this commit, the downloading of the source files has been brough into
the code box as a 'wget' command. This removes the need for a browser over
all, making it more convenient, and also avoiding the FTP problem (wget
supports FTP and has no plans to remove it).

This was reported by Fernando Buitrago.