Book: expanded the testing examples in forking tutorial
Until now, the only testing scenario in the forking tutorial was the
'./developer-build' script and a separatre mention of 'make pdf'. But
Gnuastro also has the 'tests/during-dev.sh' script which is very useful in
developments. However, adding a description of that script would have made
the box very large and unfriendly.

With this commit the discussion on the various testing methods has been
blended into the text and each are clearly separated as a Texinfo "table"
item.