MakeCatalog: multi-word options with hyphen for readability
Until now, almost all of MakeCatalog's options that contained more than one
word were simply stuck after each other. For example '--areaarcsec2' or
'--halfmaxarea' this was very hard to read and write; therefore increasing
the possibility of miss-typing debugging the source of the problem!

With this commit, a hyphen has been placed in the names of all such options
in MakeCatalog (the full list is available in the changed 'NEWS' file of
this commit). In the process, the following changes have also been made:

 - The number of MakeCatalog's measurements have become VERY long and
   having them in a single section of the book was making it increasingly
   hard to navigate and find the measurement you needed. Therefore with
   this commit, the "MakeCatalog measurements" section was broken into
   multiple sub-sections that group the various measurements by context.

 - The '--axis-ratio' and '--position-angle' options of three installed
   scripts were also updated:
   * astscript-radial-profile
   * astscript-psf-stamp
   * astscript-psf-unite

 - Most short options of MakeCatalog's measurements have been removed to
   avoid interfering with the alphabetical sorting of '--help'.