astscript-zeropoint: the name of some options are modified
Until now, some options were not clear. For instance, with '--reference'
option it was difficult realize this is for images reference or catalog
reference. On the other hand, '--catalog' option belonged to the reference
catalog but it was not clear too.

With this commit, the name of this options changed to the '--refimgs' and
'--refcat' and now they clear for users.