Book: extended PSF tutorial not using ls to count/use files
Until now, in the extended PSF tutorial, we were using 'ls' to get the list
of files for the inner and outer parts of the PSF. However, in [1], Alex
Borlaff reported that in some systems the behavior of 'ls' can cause
problems (by default it behaves like 'ls -1' in GNU Coreutils; printing
each file on a separate line).

With this commit, to avoid this situation for future users, we have removed
this style of 'ls' usage (fortunately the 'imgs' variable was not necessary
in other places of the tutorial, otherwise, we should have used a more
complicated fix).

This problem was reported by Alejandro Serrano Borlaff.

[1] https://lists.gnu.org/archive/html/bug-gnuastro/2023-01/msg00007.html