MakeCatalog: --checkuplim now returns label's random tile center
Until now, when MakeCatalog was called with '--checkuplim', it would return
the position of the first pixel of the random positioning of the label's
tile. Estimating the width and height of the object (using '--min-x',
'--max-x', '--min-y' and '--max-y') was annoying and would usually be
forgotten usually (most commonly by myself!).

As a result, when users would use these positions to place profiles (for
example in upper-limit measurements of elliptical profiles) and wouldn't
correct for the width, the randomly placed objects could fall over a
detection!

With this commit, this option now returns the center of the label's tile
instead of the "first" pixel.

This issue was discovered after a discussion with Teet Kuutma (initially I
thought it is a bug because I had forgot that the reported position is the
first pixel, not the center!)..