Library (txt.h): Not printing + infront of positive numbers
Until now, when printing any sort of number as plain text, Gnuastro's
'txt.c' library would place a '+' infront of positive numbers. This was
done so in multi-row values the positive and negative numbers don't appear
as a zig-zag when looking at the digits, like below:

   Bad         Good
   ---         -----
   1.23        +1.23
   -4.56       -4.56
   7.99        +7.99
   -1.89       -1.89

After this change was implemented Zohreh Ghaffari discovered that programs
like TOPCAT have problems with the '+' and Sepideh Eskandarlou pointed out
that the '+' is causing many complications in scripts when positive
integers were used as filenames. In both cases, they were forced to remove
the '+' with a 'sed' command before using the number.

With this commit, to have the best of both worlds (a clear visual reading
of the numbers with the first digits placed under each other without the
extra '+'), we now put a ' ' instead of the '+'. So the output above looks
like this:

   Bad         Good
   ---         -----
   1.23         1.23
   -4.56       -4.56
   7.99         7.99
   -1.89       -1.89

Ultimately this was done with a correction in the 'printf' call to define
the formats. However, it also required the following changes:

 - During tests suggested by Sepideh Eskandrlou, we discovered that
   'table_bring_to_top' was not setting the 'array' pointer of an empty
   column to NULL!

 - During the tests we noticed that the two 'table_sort' and
   'table_select_by_position' functions of 'bin/table/table.c' were only
   checking one of the ways to have an empty table (either size of zero or
   array of NULL). But to be generic it is good to check both.

 - Since we rely on many 'printf' features that may not portable in many
   non-GNU operating systems, the Gnulib 'stdio' is now also bootstrapped
   with Gnuastro.

 - Defining the new 'gal_statistics_has_negative' function to identify if
   the column has a negative value (and select the format above).

 - The 'tests/during-dev.sh' script would not rebuild the programs before
   testing the installed scripts. This would make it hard to test the
   program causing a crash within the installed scripts. From now on, all
   executable programs in the build directory are deleted and re-built
   while testing scripts.

This issue was raised by Zohreh Ghaffari and Sepideh Eskandarlou.