Library (arithmetic): new isblanknot operator
Until now, to select non-blank pixels, we would need to use these two
operators 'isblank not' in sequence. This requires one intermediate array
to be allocated and is also annoying to type.

With this commit, the arithmetic library now has the 'isnotblank'
operator. This will directly check for non-blank pixels without needing an
intermediate block in memory and is also easier to see write and read in
the command-line or a script.

This was suggested by Sepideh Eskandarlou.