Book: some parts of the tutorial modified
Until now, some parts of the tutorial used the 'Ds9' for opening the
images. While, in Gnuastro we have 'astscript-fits-view' for opening the
fits files. On the other hand, some parts needed to changed.

With this commit, insted of 'Ds9' we used 'astscript-fits-view' for opening
the astronomicla images and some parts are changed.