Bootstarmping: make a directroy for cloning the Gnulib
Until now, the variable of 'DEVDIR' define for cloning the Gnulib in
that. In the GNU Portability Library (Gnulib), section about how define
this variable was explained. But did not mentioned to make this directory
(because if the user install their OS recentely, they do not have this
directory), next command was 'cd' to this directory.

With this comment, how making this directory is added to other commands and
now it works properly for all the users.