psf-scale-factor: correcting a bug in the center position of the PSF
Until this commit, the center of the PSF was not properly computed. It was
0.5-pixel added while it should be 1-pixel added. I found this problem
while constructing the PSF of the T80 (OAJ).

With this commit I changed the 0.5 to 1.0 pixel so the PSF is properly
centered. I tested this change with a mock profile. The factor by which it
was necessary to scale a mock PSF to match with itself at a given radius
was < 1.0 (0.95 or so). Now, with the correction, it is almost 1.0
(0.99999999).