psf-scale-factor: providing a nan value as output instead of crashing
Until this commit, this script crashes when the scaling factor could not be
computed. But in some scenarios it is good to not crash but continue by
providing a nan value as output.

With this commit, this has been done. Now the script provides a nan value
when the scaling factor is not computed. This situation happens when there
is not overlap with the input image and the center coordinates provided.

This fixes task #16353: https://savannah.gnu.org/task/?16353