Tests (psf-select-stars): not keeping temporary directory
Until now, the test for 'astscript-psf-select-stars' was being executed
with the '--keeptmp' option! As a result, when 'make distcheck' was run, it
would complain about undeleted files!

With this commit, by removing the '--keeptmp' option from the test of this
script the problem was solved.

This bug was reported by Siyang He as part of task #16114.

This fixes bug #63969.