Library (pool.h): new library and arithmetic operator
Until now, there was no operator in Gnuastro for "pooling" in image
analysis. The closest operation was scaling (through warp), but that didn't
allow diffferent statistical criteria, only sum.

With this commit, the Gnuastro library now has a 'pool.h' set of functions
that can also be called within the Arithmetic program.