Book: LaTeX externalization command with no line breaks
Until now, the LaTeX (TiKZ package) externalization command (a simple shell
command) was broken into multiple lines in 'doc/plotsrc/all.tex'. On some
systems, this would result in the non-creation of the final EPS files for
each figure that we draw in TiKZ (the broken lines were not interpreted in
the "externalization" command).

With this commit, the full externalization command has been brought into
one line. Also, instead of ';' between the commands, we are now using '&&'
so it doesn' continue if one command fails.

This issue was reported by Bharat Bhandari.