Book: improved description of MakeCatalog's --clumps-sum column
Until now, the description of the '--clumps-sum' column was ambiguous. In
other words, it did not clarify that if the rivers were subtracted or not
(in the clumps catalog, did it correspond to the sum of '--sum' or
'--sum-no-river'?)

With this commit, the description has been edited to be clear about this
and avoid future confusion.

This issue was raised by Leslie Hunt and Paola Dimauro.