astscript-zeropoint: added rule to build man page
Until now, there was no rule to build a man-page for the newly added Zero
point estimation script.

With this commit, the necessary rule has been added in
'bin/script/Makefile.am'.

Also, I noticed that in some parts of the book, we are writing "zero point"
as "zeropoint" (without a space!). This has also been corrected with this
commit.