Book: direct link to SVO database of filters
Until now, we had put the top page of the SVO service while talking about
the filter data base. However, the exact data base of filters page from
that top page is not straightforward!

With this commit, the direct link to the page containing the filter data
base is given.