MakeCatalog: no warning of small --upnum when --quiet is called
Until now, the warning message for a small value to '--upnum' in
MakeCatalog would not be suppressed when the '--quiet' option was called!
Even though, the message clearly stated that it will not be printed if
called with '--quiet'.

With this commit, a check for the calling of '--quiet' has been added
before the warning message is printed.

This bug was reported by Teet Kuutma.

This fixes bug #64076.