Book: new section for shell tips
Until now, we didn't have any dedicated place in the book for generic tips
that can be useful in many programs. Such tips are usually scattered in the
tutorials or documentation of the specific programs.

With this commit, a new sub-section has been added for this under the
"Common program behavior" chapter. It is called "Shell tips" and one tip
has already been added with this commit.

This tip was written after a nice discussion with Peter Tueben.